refactor: create the `slot_assigned.rs`
